requirements.txt
azure-eventhub
azure-cognitiveservices-speech
python-dotenv
argparse
requests